Add FreeCounter component to the project
This commit adds a new file called "free-counter.tsx" which contains the implementation of the FreeCounter component. The component renders a card with information about the number of free generations available and a progress bar indicating the usage. If the user is not using the pro version, an upgrade button is displayed.